pulley/simplexity20241206/pyproject.toml
[tool.poetry]
name = "simplexity"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
ell-ai = {extras = ["all"], version = "^0.0.14"}
beautifulsoup4 = "^4.12.3"
anthropic = "0.34.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"